+++ examples/chat-rn/src/invite.tsx
@@ -1,12 +1,6 @@
+import { useAcceptInvite } from "jazz-react-native";
 import React, { Text } from "react-native";
-import { useAcceptInvite } from "./jazz";
 import { Chat } from "./schema";
-
-type ChatScreenParams = {
-  valueHint?: string;
-  valueID?: string;
-  inviteSecret?: string;
-};
 
 export default function HandleInviteScreen({
   navigation,