+++ examples/reactions/src/main.tsx
@@ -1,24 +1,20 @@
-import { DemoAuthBasicUI, createJazzReactApp, useDemoAuth } from "jazz-react";
+import { DemoAuthBasicUI, JazzProvider, useDemoAuth } from "jazz-react";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
-
-const Jazz = createJazzReactApp();
-
-export const { useAccount, useCoState } = Jazz;
 
 function JazzAndAuth({ children }: { children: React.ReactNode }) {
   const [auth, authState] = useDemoAuth();
 
   return (
     <>
-      <Jazz.Provider
+      <JazzProvider
         auth={auth}
         peer="wss://cloud.jazz.tools/?key=reactions-example@garden.co"
       >
         {children}
-      </Jazz.Provider>
+      </JazzProvider>
 
       {authState.state !== "signedIn" && (
         <DemoAuthBasicUI appName="Reactions" state={authState} />