+++ examples/chat-vue/CHANGELOG.md
@@ -1,5 +1,15 @@
 # chat-vue
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [9dd8d95]
+- Updated dependencies [8eda792]
+  - jazz-vue@0.9.0
+  - jazz-tools@0.9.0
+  - jazz-browser@0.9.0
+
 ## 0.0.30
 
 ### Patch Changes