+++ examples/reactions/CHANGELOG.md
@@ -1,5 +1,15 @@
 # reactions
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [956a4d1]
+- Updated dependencies [8eda792]
+  - jazz-react@0.9.0
+  - jazz-tools@0.9.0
+  - jazz-browser-media-images@0.9.0
+
 ## 0.0.22
 
 ### Patch Changes