+++ packages/jazz-react-auth-clerk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # jazz-browser-media-images
 
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies [8eda792]
+- Updated dependencies [956a4d1]
+- Updated dependencies [8eda792]
+- Updated dependencies [1ef3226]
+  - cojson@0.9.0
+  - jazz-react@0.9.0
+  - jazz-tools@0.9.0
+  - jazz-browser-auth-clerk@0.9.0
+
 ## 0.8.51
 
 ### Patch Changes