+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,17 @@
 # jazz-run
 
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies [8eda792]
+- Updated dependencies [8eda792]
+- Updated dependencies [1ef3226]
+  - cojson@0.9.0
+  - jazz-tools@0.9.0
+  - cojson-storage-sqlite@0.9.0
+  - cojson-transport-ws@0.9.0
+
 ## 0.8.51
 
 ### Patch Changes