+++ packages/jazz-svelte/CHANGELOG.md
@@ -1,5 +1,22 @@
 # jazz-svelte
 
+## 0.9.0
+
+### Minor Changes
+
+- 9dd8d95: Change the way the JazzProvider is created and make the API available as top-level imports.
+
+  This is a breaking change.
+
+### Patch Changes
+
+- Updated dependencies [8eda792]
+- Updated dependencies [8eda792]
+- Updated dependencies [1ef3226]
+  - cojson@0.9.0
+  - jazz-tools@0.9.0
+  - jazz-browser@0.9.0
+
 ## 0.8.51
 
 ### Patch Changes